Moved some visitors to cpp files.
