+++ include/ShaderAST/Visitors/SimplifyStatements.hpp
@@ -1,8 +1,8 @@
 /*
 See LICENSE file in root folder
 */
-#ifndef ___SDW_StmtSimplifier_H___
-#define ___SDW_StmtSimplifier_H___
+#ifndef ___SDW_SimplifyStatements_H___
+#define ___SDW_SimplifyStatements_H___
 #pragma once
 
 #include "ShaderAST/Visitors/CloneStmt.hpp"
